start tightening scheme usage in oc
