+++ cmd/oc/oc.go
@@ -10,16 +10,15 @@ import (
 	"k8s.io/apiserver/pkg/util/logs"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 
+	"github.com/openshift/api"
+	"github.com/openshift/api/authorization"
+	"github.com/openshift/api/quota"
 	"github.com/openshift/library-go/pkg/serviceability"
+	"github.com/openshift/origin/pkg/api/install"
+	"github.com/openshift/origin/pkg/api/legacy"
 	"github.com/openshift/origin/pkg/oc/cli"
 	"github.com/openshift/origin/pkg/version"
-
-	// install all APIs
-	apiinstall "github.com/openshift/origin/pkg/api/install"
-	_ "k8s.io/kubernetes/pkg/apis/autoscaling/install"
-	_ "k8s.io/kubernetes/pkg/apis/batch/install"
-	_ "k8s.io/kubernetes/pkg/apis/core/install"
-	_ "k8s.io/kubernetes/pkg/apis/extensions/install"
+	"k8s.io/kubernetes/pkg/api/legacyscheme"
 )
 
 func main() {
@@ -33,7 +32,17 @@ func main() {
 		runtime.GOMAXPROCS(runtime.NumCPU())
 	}
 
-	apiinstall.InstallAll(scheme.Scheme)
+	// the kubectl scheme expects to have all the recognizable external types it needs to consume.  Install those here.
+	api.Install(scheme.Scheme)
+	legacy.InstallExternalLegacyAll(scheme.Scheme)
+	// TODO fix up the install for the "all types"
+	authorization.Install(scheme.Scheme)
+	quota.Install(scheme.Scheme)
+
+	// the legacyscheme is used in kubectl and expects to have the internal types registered.  Explicitly wire our types here.
+	// this does
+	install.InstallInternalOpenShift(legacyscheme.Scheme)
+	legacy.InstallInternalLegacyAll(scheme.Scheme)
 
 	basename := filepath.Base(os.Args[0])
 	command := cli.CommandFor(basename)
